% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-legacy.R
\name{brvm_share_index_info}
\alias{brvm_share_index_info}
\title{Récupérer les informations sur les actions et indices de la BRVM}
\usage{
brvm_share_index_info()
}
\value{
Un data.frame contenant les colonnes suivantes :
\itemize{
\item \strong{Type}: Catégorie de l'élément (Share, Index ou Nothing).
\item \strong{Ticker}: Le symbole boursier (ex: SNTS).
\item \strong{Description}: Le nom complet de l'entreprise ou de l'indice.
\item \strong{Country.Code}: Le code pays en majuscules (ex: SN, CI).
\item \strong{Ticker_fullname}: La valeur brute combinée (ex: SNTS.SN).
}
}
\description{
Cette fonction extrait la liste des actions et des indices boursiers disponibles
sur Sika Finance. Elle traite les données brutes pour séparer les tickers,
les codes pays et les descriptions.
}
\details{
La fonction utilise le web scraping pour récupérer le contenu de la balise
sélection \code{dpShares} sur la page d'accueil de Sika Finance. Elle catégorise
ensuite chaque entrée comme "Share" (Action), "Index" (Indice) ou "Nothing".
}
\examples{
\dontrun{
 df_brvm <- brvm_share_index_info()
 head(df_brvm)
}
}
