% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvc-legacy-ticker.R
\name{bvc_market_index_info}
\alias{bvc_market_index_info}
\title{Recuperer les donnees detaillees de tous les indices de la Bourse de Casablanca}
\usage{
bvc_market_index_info(formatted = TRUE)
}
\arguments{
\item{formatted}{Logical. Indique si les colonnes de texte doivent etre formatees.}
}
\value{
Un \code{data.frame} exhaustif de l'etat des indices.
}
\description{
Version etendue de la recuperation des indices retournant l'integralite des
colonnes disponibles (valeur, plus haut, plus bas, diviseur, etc.), triee
par capitalisation decroissante.
}
