% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-legacy.R
\name{.GET_tickers_BRVM}
\alias{.GET_tickers_BRVM}
\title{Mettre à jour l'objet Market avec les tickers de la BRVM}
\usage{
.GET_tickers_BRVM()
}
\value{
Un objet de classe \code{MARKET} (tel que défini par \code{CREATE_ALL_MARKETS})
contenant les listes d'actifs mises à jour. En cas d'échec de connexion,
affiche un message d'erreur.
}
\description{
Cette fonction interne récupère les données de la BRVM via \code{brvm_share_index_info()},
les filtre pour exclure les tickers non pertinents (comme ceux de Sika Finance),
et organise les résultats dans une structure d'objet Market dédiée.
}
\details{
La fonction effectue les étapes suivantes :
\enumerate{
\item Récupération des données brutes.
\item Tri par code pays.
\item Filtrage des indices commençant par "BRVM".
\item Filtrage des actions (en excluant les préfixes "BRVM" et "SIKA").
\item Mise à jour des slots de l'objet S4 (Indexes, Shares, List, etc.).
}
}
\note{
Cette fonction est destinée à un usage interne. Elle utilise un bloc \code{tryCatch}
pour gérer les interruptions de connexion internet.
}
\keyword{internal}
