% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvc-legacy.R
\name{.GET_tickers_BVC}
\alias{.GET_tickers_BVC}
\title{Mettre a jour l'objet Market avec les tickers de la BVC}
\usage{
.GET_tickers_BVC()
}
\value{
Un objet de classe \code{MARKET} (BVC_MARKET) avec les slots \code{Indexes},
\code{Shares}, \code{List}, et \code{Ticker_full_name} mis a jour. Retourne un message
d'erreur en cas de probleme de connexion.
}
\description{
Cette fonction interne recupere les donnees de la Bourse de Casablanca (BVC) via
\code{bvc_share_index_info()}, trie les actifs par ordre alphabetique et
structure les informations dans un objet de classe MARKET.
}
\details{
Le processus se deroule en trois etapes cles :
\enumerate{
\item Appel de la fonction de base pour obtenir les actions et indices marocains.
\item Separation des donnees selon leur type ("Share" ou "Index") pour alimenter les slots S4.
\item Compilation des listes de noms et de tickers complets pour les recherches ulterieures.
}
}
\note{
Comme son nom l'indique par le prefixe ".", cette fonction est privee et
destinee a etre utilisee a l'interieur du package.
}
\keyword{internal}
