% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvc-legacy-ticker.R
\name{get_build_cache}
\alias{get_build_cache}
\title{Internal cache for build ID}
\usage{
get_build_cache()
}
\value{
Environment with cached build data.
}
\description{
This internal function creates and returns a cache environment that stores
the build ID and its timestamp. It is used internally by \code{get_build_id}
to reduce API calls.
}
\keyword{internal}
