% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvc-legacy-ticker.R
\name{get_indices_list_with_capitalization}
\alias{get_indices_list_with_capitalization}
\title{Liste des indices avec code et capitalisation triee par capitalisation decroissante}
\usage{
get_indices_list_with_capitalization(formatted = TRUE)
}
\arguments{
\item{formatted}{Logical. Indique si les valeurs numeriques doivent etre formatees. Par defaut TRUE.}
}
\value{
Un \code{data.frame} avec les colonnes \code{Indice, Code_Index, Capitalisation (MAD), URL_Complete, Categorie}.
}
\description{
Liste des indices avec code et capitalisation triee par capitalisation decroissante
}
\examples{
\dontrun{
indices <- get_indices_list_with_capitalization()
}
}
