% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse-legacy.R
\name{gse_share_info}
\alias{gse_share_info}
\title{Récupérer les indices de la Bourse de Casablanca}
\usage{
gse_share_info(table_id = "34")
}
\arguments{
\item{table_id}{Table ID.}
}
\value{
Un dataframe contenant :
\itemize{
\item \code{title} : catégorie de l'indice (Main indices, Sector Indices, etc.)
\item \code{index} : nom de l'indice
\item \code{index_code} : code extrait de l'URL (ex: MASI, FCSELIQ)
\item \code{field_index_value} : valeur actuelle de l'indice
\item \code{field_var_year} : variation annuelle (\%)
\item \code{field_var_veille} : variation journalière (\%)
\item \code{field_market_capitalisation} : capitalisation
\item \code{field_transact_time} : date/heure de la dernière transaction
\item autres variables associées aux indices
}
}
\description{
Cette fonction interroge l'API de la Bourse de Casablanca afin de récupérer
l'ensemble des indices (principaux, sectoriels, devises, etc.) sous forme
de dataframe structuré.
}
\details{
Les données sont organisées par catégorie (title) et contiennent des
informations telles que la valeur de l'indice, la capitalisation,
les variations et les bornes (haut/bas).

La fonction :
\enumerate{
\item Effectue une requête HTTP vers l'API BVC
\item Parse le JSON retourné
\item Aplati la structure imbriquée (\code{data -> items})
\item Ajoute la catégorie (\code{title})
\item Extrait le code de l'indice depuis \code{index_url}
\item Remplace les valeurs NULL par NA
\item Convertit les colonnes numériques et dates
}
}
\examples{
\dontrun{
df_indices <- get_bvc_index()
head(df_indices)
}

}
