% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-normality-test.R
\name{.normality_test_core}
\alias{.normality_test_core}
\alias{normality_test}
\alias{normality_test,numeric-method}
\alias{normality_test,ts-method}
\title{Normality Tests for Univariate Data}
\usage{
.normality_test_core(x, type.test)

normality_test(x, type.test = "ALL")

\S4method{normality_test}{numeric}(x, type.test = "ALL")

\S4method{normality_test}{ts}(x, type.test = "ALL")
}
\arguments{
\item{x}{A numeric vector or time series object.}

\item{type.test}{A character vector specifying which tests to perform.
Possible values include:
\code{"Anderson-Darling"}, \code{"Shapiro-Wilk"}, \code{"Jarque Bera"},
\code{"Cramer-von Mises"}, \code{"Shapiro-Francia"},
\code{"Lilliefors (Kolmogorov-Smirnov)"}, \code{"Pearson chi-square"},
\code{"Agostino"}, or \code{"ALL"} (default).}
}
\value{
A named list of p-values corresponding to each selected test.
Returns \code{NA} if the input data has no variability (constant values).
}
\description{
Performs a set of statistical tests to assess the normality of a numeric
vector or time series. This function is fully generic and can be applied
to any type of data (financial, economic, scientific, etc.), regardless
of the underlying market.
}
\details{
The function evaluates whether the input data follows a normal distribution
using one or multiple statistical tests. If \code{type.test = "ALL"},
all available tests are automatically performed.

The following tests are supported:
\itemize{
\item Anderson-Darling
\item Shapiro-Wilk
\item Jarque-Bera
\item Cramer-von Mises
\item Shapiro-Francia
\item Lilliefors (Kolmogorov-Smirnov)
\item Pearson Chi-square
\item D'Agostino
}

Missing values are automatically removed prior to computation.
Each test returns a p-value indicating whether the null hypothesis
of normality can be rejected.
}
\note{
This function is market-agnostic and can be used across all datasets
handled within the AfriMarkets ecosystem.
}
\examples{
\donttest{
x <- rnorm(100)

# Run all tests
normality_test(x)

# Run selected tests
normality_test(x, c("shapiro-wilk", "jarque bera"))
}

}
\seealso{
\code{\link{stationarity_test}}

Other statistical_tests: 
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE \cr
Olabiyi Aurel Geoffroy ODJO
}
\concept{statistical_tests}
